Урок 3. Домашняя работа. Прогресс бар загрузки данных. Добавлена пагинация к адаптеру. Загрузка данных с интернета через ретрофит. Кэширование в рум.

+++ app/src/main/java/ru/zfix27r/movies/data/local/entity/FilmEntity.kt
@@ -12,7 +12,7 @@ import ru.zfix27r.movies.domain.model.FilmResModel
 data class FilmEntity(
     @PrimaryKey(autoGenerate = false)
     @ColumnInfo(name = KINOPOISK_ID)
-    val kinopoiskId: Long,
+    val kinopoiskId: Int,
 
     @ColumnInfo(name = NAME_RU)
     val nameRu: String,
